matrix stress test using nursery
